# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qwizzy API Documentation",
        "description": "Documentation de l'API Qwizzy pour la gestion des quiz",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/question-types": {
            "get": {
                "tags": [
                    "QuestionType"
                ],
                "summary": "Obtenir la liste des types de questions",
                "description": "Retourne la liste paginée des types de questions",
                "operationId": "getQuestionTypesList",
                "parameters": [
                    {
                        "name": "current_sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "current_sort_dir",
                        "in": "query",
                        "description": "Direction du tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des types de questions récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "questionType": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "QuestionType"
                ],
                "summary": "Créer un nouveau type de question",
                "description": "Crée un nouveau type de question",
                "operationId": "storeQuestionType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 200,
                                        "example": "QCM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type de question créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/question-types/{id}": {
            "put": {
                "tags": [
                    "QuestionType"
                ],
                "summary": "Mettre à jour un type de question",
                "description": "Met à jour un type de question existant",
                "operationId": "updateQuestionType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type de question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 200,
                                        "example": "QCM modifié"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Type de question modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "404": {
                        "description": "Type de question non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "QuestionType"
                ],
                "summary": "Supprimer un type de question",
                "description": "Supprime un type de question existant",
                "operationId": "deleteQuestionType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type de question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type de question supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Type de question non trouvé"
                    }
                }
            }
        },
...
    },
...
}